Fix FP8 output dtype: ensure float32 before opus_writer.append_pcm
FP8 mode runs Mimi in FP16, so decode output is float16 tensor.
sphn.OpusStreamWriter.append_pcm() requires float32 numpy array.
Without .float(), crashes with: 'ndarray' cannot be converted to 'PyArray<T, D>'

This was the REAL audio output blocker — PersonaPlex was generating
audio but crashing on the Opus encoding step.

+++ moshi/moshi/server.py
@@ -346,7 +346,7 @@ class ServerState:
                         _token_count += 1
                         assert tokens.shape[1] == self.lm_gen.lm_model.dep_q + 1
                         main_pcm = self.mimi.decode(tokens[:, 1:9])
-                        pcm_cpu = main_pcm[0, 0].cpu()
+                        pcm_cpu = main_pcm[0, 0].float().cpu()  # Ensure float32 (FP8/FP16 Mimi outputs float16)
                         opus_writer.append_pcm(pcm_cpu.numpy())
                         _audio_out_count += 1
                         if _diag and _audio_out_count <= 3:
